+++ Sources/EudiWalletKit/Extensions.swift
@@ -38,6 +38,16 @@ extension Bundle {
 	}
 }
 
+extension FileManager {
+	public static func getCachesDirectory() throws -> URL {
+			let paths = FileManager.default.urls(for: .cachesDirectory, in: .userDomainMask)
+			guard paths.count > 0 else {
+				throw WalletError(description: "No downloads directory found")
+			}
+			return paths[0]
+	}
+}
+
 extension WalletStorage.Document {
 	public var authorizePresentationUrl: String? {
 		guard status == .pending, let model = try? JSONDecoder().decode(PendingIssuanceModel.self, from: data), case .presentation_request_url(let urlString) = model.pendingReason else { return nil	}